types.ts in api folder and edit profile
