+++ src/app/(routes)/recruiter/profile/page.tsx
@@ -5,11 +5,13 @@ import RecruiterProfile from "@/components/Recruiters/profile";
 import { fetchProfile } from "@/helpers/recruiter/api";
 import Cookies from "js-cookie";
 import loadingImg from "@/../public/loadingSpinner.svg";
-import { ProfileFC } from "@/helpers/recruiter/api";
+import { ProfileFC } from "@/helpers/recruiter/types";
+import EditProfilePage from "@/components/Recruiters/editProfile";
 
-const Profile = ({ params }: { params: { RecruiterId: string } }) => {
+const Profile = () => {
   const [data, setData] = useState<ProfileFC>();
   const [loading, setLoading] = useState(true);
+  const [editMode, setEditMode] = useState(false);
 
   useEffect(() => {
     const profileData = async () => {
@@ -23,7 +25,15 @@ const Profile = ({ params }: { params: { RecruiterId: string } }) => {
   return (
     <div className="h-screen flex justify-center">
       {loading && <img src={loadingImg.src} width={200} />}
-      {data && <RecruiterProfile profile={data} />}
+      {data &&
+        (editMode ? (
+          <EditProfilePage data={data} />
+        ) : (
+          <RecruiterProfile
+            profile={data}
+            setEdit={() => setEditMode(!editMode)}
+          />
+        ))}
     </div>
   );
 };
