+++ src/helpers/recruiter/api.ts
@@ -1,5 +1,14 @@
-// pages/api/fetchPdf.js
 import axios from "axios";
+import {
+  JobDetailFC,
+  JobsFC,
+  JAFdetailsFC,
+  ProfileFC,
+  updateProfileFC,
+  EventFC,
+  ApplicationFC,
+  SalaryFC,
+} from "./types";
 
 const redirect = () => {};
 
@@ -9,34 +18,6 @@ process.env.NODE_TLS_REJECT_UNAUTHORIZED = "0";
 const url = (NextUrl: string) => {
   return `${baseUrl}/api/v1${NextUrl}`;
 };
-
-export interface ProfileFC {
-  id: string;
-  designation: string;
-  landline: string;
-  user: {
-    id: string;
-    email: string;
-    name: string;
-    contact: string;
-  };
-  company: {
-    name: string;
-    domains: [string];
-    category: string;
-    address: {
-      city: string;
-      line1: string;
-      line2: string;
-      state: string;
-      country: string;
-      zipCode: string;
-    };
-    size: number;
-    yearOfEstablishment: string;
-    socialMediaLink: string;
-  };
-}
 
 export const fetchProfile = async (accessToken: string | undefined) => {
   if (!accessToken || accessToken === undefined) {
@@ -52,33 +33,6 @@ export const fetchProfile = async (accessToken: string | undefined) => {
   const json: ProfileFC = await res.json();
   return json;
 };
-
-export interface updateProfileFC {
-  designation?: string;
-  landline?: string;
-  company?: {
-    name?: string;
-    website?: string;
-    domains?: any[];
-    category?: string;
-    address?: {
-      line1: string;
-      line2: string;
-      city: string;
-      state: string;
-      country: string;
-    };
-    size?: number;
-    yearOfEstablishment?: string;
-    annualTurnover?: string;
-    socialMediaLink?: string;
-  };
-  user?: {
-    name: string;
-    email: string;
-    contact: string;
-  };
-}
 
 export const patchProfile = async (
   accessToken: string | undefined,
@@ -99,22 +53,6 @@ export const patchProfile = async (
   return res.ok;
 };
 
-export interface JobsFC {
-  id: string;
-  role: string;
-  active: boolean;
-  currentStatus: string;
-  season: {
-    id: string;
-    year: string;
-    type: string;
-  };
-  company: {
-    id: string;
-    name: string;
-  };
-}
-
 export const getJobs = async (accessToken: string | undefined) => {
   if (!accessToken || accessToken === undefined) {
     redirect();
@@ -129,100 +67,6 @@ export const getJobs = async (accessToken: string | undefined) => {
   const json: JobsFC[] = await res.json();
   return json;
 };
-
-export interface JobDetailFC {
-  id: string;
-  role: string;
-  active: boolean;
-  currentStatus: string;
-  season: {
-    id: string;
-    year: string;
-    type: string;
-  };
-  company: {
-    id: string;
-    name: string;
-  };
-  selectionProcedure: {
-    selectionMode: string;
-    shortlistFromResume: true;
-    groupDiscussion: true;
-    tests: [
-      {
-        type: string;
-        duration: number;
-      },
-    ];
-    interviews: [
-      {
-        type: string;
-        duration: number;
-      },
-    ];
-    requirements: {
-      numberOfMembers: number;
-      numberOfRooms: number;
-      otherRequirements: string;
-    };
-    others: string;
-  };
-  description: string;
-  attachment: string;
-  skills: string;
-  offerLetterReleaseDate: string;
-  joiningDate: string;
-  location: string;
-  noOfVacancies: number;
-  duration: number;
-  feedback: string;
-  jobCoordinators: [
-    {
-      id: string;
-      role: string;
-      tpcMember: {
-        id: string;
-        department: string;
-        role: string;
-        user: {
-          id: string;
-          email: string;
-          name: string;
-          contact: string;
-        };
-      };
-    },
-  ];
-  events: [
-    {
-      id: string;
-      roundNumber: number;
-      type: string;
-      metadata: string;
-      startDateTime: string;
-      endDateTime: string;
-      visibleToRecruiter: true;
-    },
-  ];
-  salaries: [SalaryFC];
-}
-
-interface SalaryFC {
-  id: string;
-  baseSalary: string;
-  takeHomeSalary: number;
-  grossSalary: number;
-  otherCompensations: number;
-  totalCTC: number;
-  salaryPeriod: string;
-  genders: [string];
-  programs: [string];
-  facultyApprovals: [string];
-  categories: [string];
-  minCPI: number;
-  tenthMarks: number;
-  twelthMarks: number;
-}
 
 export const getJobDetail = async (
   accessToken: string | undefined,
@@ -241,42 +85,6 @@ export const getJobDetail = async (
   const json: JobDetailFC = await res.json();
   return json;
 };
-
-export interface ApplicationFC {
-  id: string;
-  student: {
-    id: string;
-    rollNo: string;
-    user: {
-      id: string;
-      name: string;
-      email: string;
-      contact: string;
-    };
-    program: {
-      id: string;
-      branch: string;
-      course: string;
-      year: string;
-      department: string;
-    };
-  };
-  resume: {
-    id: string;
-    filepath: string;
-    verified: boolean;
-  };
-}
-
-export interface EventFC {
-  id: string;
-  roundNumber: number;
-  type: string;
-  metadata: string;
-  startDateTime: string;
-  endDateTime: string;
-  applications?: [ApplicationFC];
-}
 
 export async function getEvent(accessToken: string | undefined, id: string) {
   if (!accessToken || accessToken === undefined) {
@@ -306,31 +114,6 @@ export async function getDomains(accessToken: string | undefined) {
   });
   const json = await res.json();
   return json.domains;
-}
-
-export interface JAFdetailsFC {
-  seasons: [
-    {
-      id: string;
-      type: string;
-      year: string;
-    },
-  ];
-  programs: [
-    {
-      id: string;
-      branch: string;
-      course: string;
-      year: string;
-      department: string;
-    },
-  ];
-  genders: [string];
-  categories: [string];
-  testTypes: [string];
-  domains: [string];
-  interviewTypes: [string];
-  countries: [string];
 }
 
 export async function getJafDetails(accessToken: string | undefined) {
